+++ src/burn/drv/pre90s/d_mrdo.cpp
@@ -206,7 +206,7 @@ static void mrdo_palette_init()
 		}
 		else pot[i] = 0;
 
-		weight[i] = 0xff * pot[i] / pot[0x0f];
+		weight[i] = (INT32)(0xff * pot[i] / pot[0x0f]);
 		if (weight[i] < 0) weight[i] = 0;
 	}
